third_party/blink/web_tests/platform/linux-chrome/external/wpt/svg-aam/role/roles-generic-expected.txt

This is a testharness.js-based test.
[FAIL] el-image
  assert_equals: <image data-testname="el-image" data-expectedrole="image" class="ex"></image> expected "image" but got "none"
Harness: the test ran to completion.
